// src/modules/post/types.tsx
export interface PostResponse {
	message: string;
	data: Post;
}

export interface PostsResponse {
	message: string;
	data: Post[];
}

export interface Post {
	id: number;
	title: string;
	slug: string;
	content: string;
	excerpt: string;
	status: string;
	createdAt: string;
}

export interface CreatePostInput {
	title: string;
	slug: string;
	content: string;
	excerpt: string;
}